- Auto-detect processor count when hashing in parallel
- Detect image files based on mime type instead of file name
- Support for MongoDB URIs (#19)
- Add support for CORS

+++ duplicate_finder.py
@@ -13,10 +13,10 @@ Usage:
 Options:
     -h, --help                Show this screen
 
-    --db=<db_path>            The location of the database. (default: ./db)
+    --db=<db_path>            The location of the database or a MongoDB URI. (default: ./db)
 
     --parallel=<num_processes> The number of parallel processes to run to hash the image
-                               files (default: 8).
+                               files (default: number of CPUs).
 
     find:
         --print               Only print duplicate files rather than displaying HTML file
@@ -29,14 +29,17 @@ Options:
 import concurrent.futures
 from contextlib import contextmanager
 import os
+import magic
+import math
 from pprint import pprint
+import psutil
 import shutil
 from subprocess import Popen, PIPE, TimeoutExpired
 from tempfile import TemporaryDirectory
 import webbrowser
-import math
 
 from flask import Flask
+from flask_cors import CORS
 import imagehash
 from jinja2 import FileSystemLoader, Environment
 from more_itertools import chunked
@@ -47,51 +50,75 @@ from termcolor import cprint
 
 TRASH = "./Trash/"
 DB_PATH = "./db"
-NUM_PROCESSES = 8
+NUM_PROCESSES = psutil.cpu_count()
 
 
 @contextmanager
-def connect_to_db():
-    if not os.path.isdir(DB_PATH):
-        os.makedirs(DB_PATH)
+def connect_to_db(db_conn_string='./db'):
+    p = None
 
-    p = Popen(['mongod', '--dbpath', DB_PATH], stdout=PIPE, stderr=PIPE)
+    # Determine db_conn_string is a mongo URI or a path
+    # If this is a URI
+    if 'mongodb://' == db_conn_string[:10]:
+        client = pymongo.MongoClient(db_conn_string)
+        cprint("Connected server...", "yellow")
+        db = client.image_database
+        images = db.images
 
-    try:
-        p.wait(timeout=2)
-        stdout, stderr = p.communicate()
-        cprint("Error starting mongod", "red")
-        cprint(stdout.decode(), "red")
-        exit()
-    except TimeoutExpired:
-        pass
+    # If this is not a URI
+    else:
+        if not os.path.isdir(db_conn_string):
+            os.makedirs(db_conn_string)
 
-    cprint("Started database...", "yellow")
-    client = pymongo.MongoClient()
-    db = client.image_database
-    images = db.images
+        p = Popen(['mongod', '--dbpath', db_conn_string], stdout=PIPE, stderr=PIPE)
+
+        try:
+            p.wait(timeout=2)
+            stdout, stderr = p.communicate()
+            cprint("Error starting mongod", "red")
+            cprint(stdout.decode(), "red")
+            exit()
+        except TimeoutExpired:
+            pass
+
+        cprint("Started database...", "yellow")
+        client = pymongo.MongoClient()
+        db = client.image_database
+        images = db.images
 
     yield images
 
     client.close()
-    cprint("Stopped database...", "yellow")
-    p.terminate()
+
+    if p is not None:
+        cprint("Stopped database...", "yellow")
+        p.terminate()
 
 
 def get_image_files(path):
+    """
+    Check path recursively for files. If any compatible file is found, it is
+    yielded with its full path.
+
+    :param path:
+    :return: yield absolute path
+    """
     def is_image(file_name):
-        file_name = file_name.lower()
-        return file_name.endswith('.jpg') or  \
-            file_name.endswith('.jpeg') or \
-            file_name.endswith('.png') or  \
-            file_name.endswith('.gif') or  \
-            file_name.endswith('.tiff')
+        # List mime types fully supported by Pillow
+        full_supported_formats = ['gif', 'jp2', 'jpeg', 'pcx', 'png', 'tiff', 'x-ms-bmp',
+                                  'x-portable-pixmap', 'x-xbitmap']
+        try:
+            mime = magic.from_file(file_name, mime=True)
+            return mime.rsplit('/', 1)[1] in full_supported_formats
+        except IndexError:
+            return False
 
     path = os.path.abspath(path)
     for root, dirs, files in os.walk(path):
         for file in files:
+            file = os.path.join(root, file)
             if is_image(file):
-                yield os.path.join(root, file)
+                yield file
 
 
 def hash_file(file):
@@ -261,6 +288,7 @@ def display_duplicates(duplicates, db):
         regex = '.*?'
 
     app = Flask(__name__)
+    CORS(app)
     app.url_map.converters['everything'] = EverythingConverter
 
     def render(duplicates, current, total):
@@ -324,7 +352,7 @@ if __name__ == '__main__':
     if args['--parallel']:
         NUM_PROCESSES = int(args['--parallel'])
 
-    with connect_to_db() as db:
+    with connect_to_db(db_conn_string=DB_PATH) as db:
         if args['add']:
             add(args['<path>'], db)
         elif args['remove']:
